# docs/contracts/openapi/deployer.openapi.json
{
  "components": {
    "schemas": {
...
      "DeploymentAccessCredentialResult": {
        "additionalProperties": false,
        "properties": {
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "layer": {
            "pattern": "^l5$",
            "title": "Layer",
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "provider": {
            "pattern": "^gcp$",
            "title": "Provider",
            "type": "string"
          },
          "schema_version": {
            "pattern": "^deployment-access-credential\\.v1$",
            "title": "Schema Version",
            "type": "string"
          },
          "username": {
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "layer",
          "provider",
          "username",
          "password",
          "issued_at"
        ],
        "title": "DeploymentAccessCredentialResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for deploying, destroying, and inspecting Digital Twin resources.",
    "title": "Digital Twin Manager API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/infrastructure/cooldown-check": {
      "get": {
        "description": "Check if redeployment is allowed (GCP Firestore 5-min cooldown).\n\n**Zero cloud costs:** Pure calculation with hardcoded 5-min limit. No cloud API calls.\n\nArgs:\n    destroyed_at: ISO timestamp from Management API\n    uses_gcp_firestore: Whether deployment uses GCP as L3-Hot provider\n\nReturns:\n    ready: True if deployment can proceed\n    remaining_seconds: Seconds until ready (0 if ready)",
        "operationId": "check_cooldown_infrastructure_cooldown_check_get",
        "parameters": [
          {
            "description": "ISO timestamp of last destroy",
            "in": "query",
            "name": "destroyed_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO timestamp of last destroy",
              "title": "Destroyed At"
            }
          },
          {
            "description": "Whether deployment uses GCP Firestore",
            "in": "query",
            "name": "uses_gcp_firestore",
            "required": false,
            "schema": {
              "default": true,
              "description": "Whether deployment uses GCP Firestore",
              "title": "Uses Gcp Firestore",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Cooldown status returned"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check GCP Firestore deployment cooldown",
        "tags": [
          "Infrastructure"
        ]
      }
    },
    "/infrastructure/deploy": {
      "post": {
        "description": "Deploys the full digital twin environment using Terraform.\n\n**Deployment process:**\n1. Validates project structure and configuration\n2. Runs `terraform init` (if needed) and `terraform apply`\n3. Deploys all configured layers based on config_providers.json\n\n**Layers deployed:**\n- **L1** (IoT Ingestion): IoT Hub/Core, Dispatcher Lambda/Function\n- **L2** (Processing): Persister, Event Checker, State Machine\n- **L3** (Storage): Hot storage (DynamoDB/CosmosDB), Cold storage (S3/Blob)\n- **L4** (Digital Twin): TwinMaker/ADT entities\n- **L5** (Visualization): Grafana dashboards\n\n**Note:** Long-running operation (2-10 minutes depending on resources).",
        "operationId": "deploy_all_infrastructure_deploy_post",
        "parameters": [
          {
            "description": "Cloud provider: aws, azure, or google",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "aws",
              "description": "Cloud provider: aws, azure, or google",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "description": "Name of the project context",
            "in": "query",
            "name": "project_name",
            "required": false,
            "schema": {
              "default": "template",
              "description": "Name of the project context",
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Operation-Package",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Operation-Package",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Deployment successful"
          },
          "400": {
            "description": "Invalid project or provider"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Deployment failed"
          }
        },
        "summary": "Deploy full digital twin environment",
        "tags": [
          "Infrastructure"
        ]
      }
    },
    "/infrastructure/deploy/stream": {
      "post": {
        "description": "Deploy with Server-Sent Events streaming.\n\nReturns an SSE stream with real-time deployment logs.",
        "operationId": "deploy_stream_infrastructure_deploy_stream_post",
        "parameters": [
          {
            "description": "Cloud provider: aws, azure, or google",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "aws",
              "description": "Cloud provider: aws, azure, or google",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "description": "Name of the project context",
            "in": "query",
            "name": "project_name",
            "required": false,
            "schema": {
              "default": "template",
              "description": "Name of the project context",
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Operation-Package",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Operation-Package",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "SSE stream of deployment logs"
          },
          "400": {
            "description": "Invalid project or provider"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Deployment failed"
          }
        },
        "summary": "Deploy with SSE streaming logs",
        "tags": [
          "Infrastructure"
        ]
      }
    },
    "/infrastructure/deployment-access/l5/credentials:rotate": {
      "post": {
        "operationId": "rotate_deployment_access_credential_infrastructure_deployment_access_l5_credentials_rotate_post",
        "parameters": [
          {
            "description": "Name of the project context",
            "in": "query",
            "name": "project_name",
            "required": true,
            "schema": {
              "description": "Name of the project context",
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Operation-Package",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Operation-Package",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentAccessCredentialResult"
                }
              }
            },
            "description": "One-time Viewer credential"
          },
          "400": {
            "description": "Invalid or expired operation package"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "description": "Bounded GKE rotation failed"
          }
        },
        "summary": "Rotate the deployed GCP Grafana Viewer credential once",
        "tags": [
          "Infrastructure"
        ]
      }
    },
    "/infrastructure/destroy": {
      "post": {
        "description": "Destroys the full digital twin environment using Terraform.\n\n**Destruction process:**\n1. Runs `terraform destroy` to remove all infrastructure\n2. Cleans up SDK-managed resources (IoT devices, Digital Twin entities)\n\n**If destruction fails for AWS TwinMaker:**\nUse `DELETE /projects/{name}/cleanup/aws-twinmaker` to manually clean entities first.\n\n**Note:** This operation cannot be undone. All data will be lost.",
        "operationId": "destroy_all_infrastructure_destroy_post",
        "parameters": [
          {
            "description": "Cloud provider: aws, azure, or google",
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "default": "aws",
              "description": "Cloud provider: aws, azure, or google",
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "description": "Name of the project context",
            "in": "query",
            "name": "project_name",
            "required": false,
            "schema": {
              "default": "template",
              "description": "Name of the project context",
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Operation-Package",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Operation-Package",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Destruction successful"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Destruction failed - may need force cleanup"
          }
        },
        "summary": "Destroy full digital twin environment",
        "tags": [
          "Infrastructure"
        ]
      }
    },
...
  },
...
}